+++ src/enums/fees.h
@@ -10,6 +10,8 @@
 #ifndef FEES_H
 #define FEES_H
 
+#include <cstdint>
+
 namespace Ark {
 namespace Crypto {
 namespace Enums {
@@ -21,15 +23,15 @@ namespace Enums {
 struct Fees
 {
     public:
-        static const unsigned int TRANSFER                      = 10000000;
-        static const unsigned int SECOND_SIGNATURE_REGISTRATION = 500000000;
-        static const unsigned int DELEGATE_REGISTRATION         = 2500000000;
-        static const unsigned int VOTE                          = 100000000;
-        static const unsigned int MULTI_SIGNATURE_REGISTRATION  = 500000000;
-        static const unsigned int IPFS                          = 0;
-        static const unsigned int TIMELOCK_TRANSFER             = 0;
-        static const unsigned int MULTI_PAYMENT                 = 0;
-        static const unsigned int DELEGATE_RESIGNATION          = 0;
+        static const uint64_t TRANSFER                      = 10000000;
+        static const uint64_t SECOND_SIGNATURE_REGISTRATION = 500000000;
+        static const uint64_t DELEGATE_REGISTRATION         = 2500000000;
+        static const uint64_t VOTE                          = 100000000;
+        static const uint64_t MULTI_SIGNATURE_REGISTRATION  = 500000000;
+        static const uint64_t IPFS                          = 0;
+        static const uint64_t TIMELOCK_TRANSFER             = 0;
+        static const uint64_t MULTI_PAYMENT                 = 0;
+        static const uint64_t DELEGATE_RESIGNATION          = 0;
 };
 /**/
 };
